agents_catalog/32-sila2-lab-automation-agent/mock_devices/requirements.txt
grpcio==1.60.0
protobuf==4.25.1
